+++ circus/client.py
@@ -5,6 +5,7 @@ import zmq
 import zmq.utils.jsonapi as json
 from zmq.eventloop.zmqstream import ZMQStream
 import tornado
+from tornado import concurrent
 
 from circus.exc import CallError
 from circus.util import DEFAULT_ENDPOINT_DEALER, get_connection, to_bytes
@@ -35,7 +36,7 @@ class AsyncCircusClient(object):
         get_connection(self.socket, endpoint, ssh_server, ssh_keyfile)
         self._timeout = timeout
         self.timeout = timeout * 1000
-        self.stream = ZMQStream(self.socket, tornado.ioloop.IOLoop.instance())
+        self.stream = ZMQStream(self.socket, tornado.ioloop.IOLoop.current())
 
     def _init_context(self, context):
         self.context = context or zmq.Context.instance()
@@ -65,12 +66,20 @@ class AsyncCircusClient(object):
             raise CallError(str(e))
 
         try:
-            yield tornado.gen.Task(self.stream.send, cmd)
+            future = concurrent.Future()
+
+            def cb(msg, status):
+                future.set_result(msg)
+            self.stream.send(cmd, callback=cb)
+            yield future
         except zmq.ZMQError as e:
             raise CallError(str(e))
 
         while True:
-            messages = yield tornado.gen.Task(self.stream.on_recv)
+            future = concurrent.Future()
+            self.stream.on_recv(future.set_result)
+            messages = yield future
+
             for message in messages:
                 try:
                     res = json.loads(message)