+++ circus/stats/collector.py
@@ -12,7 +12,7 @@ class BaseStatsCollector(ioloop.PeriodicCallback):
 
     def __init__(self, streamer, name, callback_time=1., io_loop=None):
         ioloop.PeriodicCallback.__init__(self, self._callback,
-                                         callback_time * 1000, io_loop)
+                                         callback_time * 1000)
         self.streamer = streamer
         self.name = name
 
@@ -110,8 +110,7 @@ class SocketStatsCollector(BaseStatsCollector):
                                                    callback_time, io_loop)
         self._rstats = defaultdict(int)
         self.sockets = [sock for sock, address, fd in self.streamer.sockets]
-        self._p = ioloop.PeriodicCallback(self._select, _LOOP_RES,
-                                          io_loop=io_loop)
+        self._p = ioloop.PeriodicCallback(self._select, _LOOP_RES)
 
     def start(self):
         self._p.start()