+++ circus/tests/test_plugin_flapping.py
@@ -1,4 +1,4 @@
-from mock import patch
+from unittest.mock import patch
 
 from circus.tests.support import TestCircus, EasyTestSuite
 from circus.plugins.flapping import Flapping