+++ examples/test.sh
@@ -1,2 +1,3 @@
 #!/bin/sh
-;sleep 1
+sleep 1
+echo "@@@"